requirements.txt
google-adk[a2a]>=1.25.0
litellm>=1.40.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.30.0
